+++ R/theme.r
@@ -156,19 +156,14 @@
 #' @param strip.switch.pad.wrap space between strips and axes when strips are
 #'   switched (`unit`)
 #'
-#' @param ... additional element specifications not part of base ggplot2. If
-#'   supplied `validate` needs to be set to `FALSE`.
+#' @param ... additional element specifications not part of base ggplot2. In general,
+#'   these should also be defined in the `element tree` argument.
 #' @param complete set this to `TRUE` if this is a complete theme, such as
 #'   the one returned by [theme_grey()]. Complete themes behave
 #'   differently when added to a ggplot object. Also, when setting
 #'   `complete = TRUE` all elements will be set to inherit from blank
 #'   elements.
 #' @param validate `TRUE` to run `validate_element()`, `FALSE` to bypass checks.
-#' @param element_tree optional addition or modification to the element tree,
-#'   which specifies the inheritance relationship of the theme elements. The element
-#'   tree should be provided as a list of named element definitions created with
-#'   [`el_def()`]. See [`el_def()`] for more details.
-#'
 #' @seealso
 #'   [+.gg()] and [%+replace%],
 #'   [element_blank()], [element_line()],
@@ -362,10 +357,9 @@ theme <- function(line,
                   strip.switch.pad.wrap,
                   ...,
                   complete = FALSE,
-                  validate = TRUE,
-                  element_tree = NULL
+                  validate = TRUE
                   ) {
-  elements <- find_args(..., complete = NULL, validate = NULL, element_tree = NULL)
+  elements <- find_args(..., complete = NULL, validate = NULL)
 
   if (!is.null(elements$axis.ticks.margin)) {
     warn("`axis.ticks.margin` is deprecated. Please set `margin` property of `axis.text` instead")
@@ -405,8 +399,7 @@ theme <- function(line,
     elements,
     class = c("theme", "gg"),
     complete = complete,
-    validate = validate,
-    element_tree = element_tree
+    validate = validate
   )
 }
 
@@ -420,24 +413,6 @@ is_theme_validate <- function(x) {
     TRUE # we validate by default
   else
     isTRUE(validate)
-}
-
-# obtain the full element tree from a theme,
-# substituting the defaults if needed
-complete_element_tree <- function(theme) {
-  element_tree <- attr(theme, "element_tree", exact = TRUE)
-
-  # we fill in the element tree first from the current default theme,
-  # and then from the internal element tree if necessary
-  # this makes it easy for extension packages to provide modified
-  # default element trees
-  defaults(
-    defaults(
-      element_tree,
-      attr(theme_get(), "element_tree", exact = TRUE)
-    ),
-    ggplot_global$element_tree
-  )
 }
 
 # Combine plot defaults with current theme to get complete theme for a plot
@@ -455,15 +430,15 @@ plot_theme <- function(x, default = theme_get()) {
     theme <- default + theme
   }
 
-  # complete the element tree and save back to the theme
-  element_tree <- complete_element_tree(theme)
-  attr(theme, "element_tree") <- element_tree
+  # if we're still missing elements relative to fallback default, fill in those
+  missing <- setdiff(names(ggplot_global$theme_default), names(theme))
+  theme[missing] <- ggplot_global$theme_default[missing]
 
   # Check that all elements have the correct class (element_text, unit, etc)
   if (is_theme_validate(theme)) {
     mapply(
       validate_element, theme, names(theme),
-      MoreArgs = list(element_tree = element_tree)
+      MoreArgs = list(element_tree = get_element_tree())
     )
   }
 
@@ -503,12 +478,6 @@ add_theme <- function(t1, t2, t2name) {
   # Only validate if both themes should be validated
   attr(t1, "validate") <-
     is_theme_validate(t1) && is_theme_validate(t2)
-
-  # Merge element trees if provided
-  attr(t1, "element_tree") <- defaults(
-    attr(t2, "element_tree", exact = TRUE),
-    attr(t1, "element_tree", exact = TRUE)
-  )
 
   t1
 }
@@ -552,13 +521,8 @@ calc_element <- function(element, theme, verbose = FALSE, skip_blank = FALSE) {
     }
   }
 
-  # Obtain the element tree and check that the element is in it
-  # If not, try to retrieve the complete element tree. This is
-  # needed for backwards compatibility and certain unit tests.
-  element_tree <- attr(theme, "element_tree", exact = TRUE)
-  if (!element %in% names(element_tree)) {
-    element_tree <- complete_element_tree(theme)
-  }
+  # Obtain the element tree
+  element_tree <- get_element_tree()
 
   # If the element is defined (and not just inherited), check that
   # it is of the class specified in element_tree
@@ -580,8 +544,8 @@ calc_element <- function(element, theme, verbose = FALSE, skip_blank = FALSE) {
       return(el_out) # no null properties, return element as is
     }
 
-    # if we have null properties, try to fill in from theme_grey()
-    el_out <- combine_elements(el_out, ggplot_global$theme_grey[[element]])
+    # if we have null properties, try to fill in from ggplot_global$theme_default
+    el_out <- combine_elements(el_out, ggplot_global$theme_default[[element]])
     nullprops <- vapply(el_out, is.null, logical(1))
     if (!any(nullprops)) {
       return(el_out) # no null properties remaining, return element