+++ tests/testthat/test-theme.r
@@ -265,13 +265,13 @@ test_that("element tree can be modified", {
   expect_error(print(p), "Theme element `blablabla` is not defined in the element hierarchy")
 
   # things work once we add a new element to the element tree
-  q <- p + theme(
+  register_theme_elements(
     element_tree = list(blablabla = el_def("element_text", "text"))
   )
-  expect_silent(print(q))
+  expect_silent(print(p))
 
   # inheritance and final calculation of novel element works
-  final_theme <- ggplot2:::plot_theme(q, theme_gray())
+  final_theme <- ggplot2:::plot_theme(p, theme_gray())
   e1 <- calc_element("blablabla", final_theme)
   e2 <- calc_element("text", final_theme)
   expect_identical(e1$family, e2$family)
@@ -279,6 +279,15 @@ test_that("element tree can be modified", {
   expect_identical(e1$size, e2$size)
   expect_identical(e1$lineheight, e2$lineheight)
   expect_identical(e1$colour, "red") # not inherited from element_text
+
+  # existing elements can be overwritten
+  ed <- el_def("element_rect", "rect")
+  register_theme_elements(
+    element_tree = list(axis.title = ed)
+  )
+  expect_identical(get_element_tree()$axis.title, ed)
+
+  reset_theme_settings(reset_current = FALSE) # revert back to defaults
 })
 
 test_that("all elements in complete themes have inherit.blank=TRUE", {
@@ -362,10 +371,9 @@ test_that("current theme can be updated with new elements", {
   expect_identical(calc_element("abcde", plot_theme(b1)), NULL)
 
   # element tree gets merged properly
-  theme_replace(
+  register_theme_elements(
     abcde = element_text(color = "blue", hjust = 0, vjust = 1),
-    element_tree = list(abcde = el_def("element_text", "text")),
-    complete = TRUE
+    element_tree = list(abcde = el_def("element_text", "text"))
   )
 
   e1 <- calc_element("abcde", plot_theme(b2))
@@ -375,6 +383,7 @@ test_that("current theme can be updated with new elements", {
   e2$vjust <- 1
   expect_identical(e1, e2)
 
+  reset_theme_settings()
   theme_set(old)
 })
 
@@ -577,7 +586,7 @@ test_that("plot titles and caption can be aligned to entire plot", {
 })
 
 test_that("provided themes explicitly define all elements", {
-  elements <- names(ggplot_global$element_tree)
+  elements <- names(.element_tree)
 
   t <- theme_all_null()
   expect_true(all(names(t) %in% elements))